flags: url

+++ email/url.c
@@ -350,7 +350,7 @@ err:
  * @retval  0 Success
  * @retval -1 Error
  */
-int url_tobuffer(struct Url *url, struct Buffer *buf, int flags)
+int url_tobuffer(struct Url *url, struct Buffer *buf, uint8_t flags)
 {
   if (!url || !buf)
     return -1;
@@ -415,7 +415,7 @@ int url_tobuffer(struct Url *url, struct Buffer *buf, int flags)
  * @retval  0 Success
  * @retval -1 Error
  */
-int url_tostring(struct Url *url, char *dest, size_t len, int flags)
+int url_tostring(struct Url *url, char *dest, size_t len, uint8_t flags)
 {
   if (!url || !dest)
     return -1;
